allow whitespace at the beginning of expressions
This commit fixes a bug in the Parser, wherein whitespace was not being allowed
at the beginning of expressions, while being allowed between expressions and at
the end of expressions, creating a strange UI inconsistency.

Now, any amount of whitespace is allowed anywhere between and around terms, and
same for expressions (by extension).