Add support for fractional numbers of seconds
`hms` now uses the Data.Scientific type to store fractional numbers of seconds
in arbitrary precision. This features significantly extends the utility of `hms`
and makes it easy to use with common command-line tools that output the duration
of audio files.

Multiple notations for entering fraction numbers of seconds are supported, and
are all usable from within the hours, minutes, or seconds field of a duration
term. E.g.,

  > hms "10.24 + 20.35"
  # 00:00:30.59

  > hms "0.5:0:0"
  # 00:30:00.0

  > hms "00:1/4:00"
  # 00:00:15.0